avocado.plugins: Split sysinfo and test_lister plugins
Put the sysinfo and test_lister plugins into separate
plugin files. Although not strictly necessary, the
neat 1 plugin - 1 source file relation makes things
organized and easy to understand.
