+++ avocado/plugins/plugins_list.py
