+++ avocado/plugins/runner.py
@@ -16,66 +16,12 @@
 Base Test Runner Plugins.
 """
 
-import os
 import sys
 
+from avocado.core import error_codes
 from avocado.plugins import plugin
-from avocado.core import data_dir
 from avocado.core import output
-from avocado.utils import path
-from avocado import sysinfo
 from avocado import job
-from avocado.settings import settings
-
-
-class TestLister(plugin.Plugin):
-
-    """
-    Implements the avocado 'list' subcommand
-    """
-
-    name = 'test_lister'
-    enabled = True
-
-    def configure(self, parser):
-        """
-        Add the subparser for the list action.
-
-        :param parser: Main test runner parser.
-        """
-        self.parser = parser.subcommands.add_parser(
-            'list',
-            help='List available test modules')
-        super(TestLister, self).configure(self.parser)
-
-    def run(self, args):
-        """
-        List available test modules.
-
-        :param args: Command line args received from the list subparser.
-        """
-        view = output.View(app_args=args, use_paginator=True)
-        base_test_dir = data_dir.get_test_dir()
-        test_files = os.listdir(base_test_dir)
-        test_dirs = []
-        blength = 0
-        for t in test_files:
-            inspector = path.PathInspector(path=t)
-            if inspector.is_python():
-                clength = len((t.split('.')[0]))
-                if clength > blength:
-                    blength = clength
-                test_dirs.append((t.split('.')[0], os.path.join(base_test_dir, t)))
-        format_string = "    %-" + str(blength) + "s %s"
-        view.notify(event="message", msg='Config file path: %s' % settings.config_path)
-        view.notify(event="minor", msg='')
-        view.notify(event="message", msg='Tests dir: %s' % base_test_dir)
-        if len(test_dirs) > 0:
-            view.notify(event="minor", msg=format_string % ('Alias', 'Path'))
-            for test_dir in test_dirs:
-                view.notify(event="minor", msg=format_string % test_dir)
-        else:
-            view.notify(event="error", msg='No tests were found on current tests dir')
 
 
 class TestRunner(plugin.Plugin):
@@ -168,16 +114,15 @@ class TestRunner(plugin.Plugin):
 
         :param args: Command line args received from the run subparser.
         """
-
+        view = output.View(app_args=args, use_paginator=True)
         if args.unique_job_id is not None:
             try:
                 int(args.unique_job_id, 16)
                 if len(args.unique_job_id) != 40:
-                    raise Exception
-            except:
-                print >> sys.stderr, \
-                    'Error: Unique Job ID needs to be a 40 digit hex number'
-                return -1
+                    raise ValueError
+            except ValueError:
+                view.notify(event='error', msg='Unique Job ID needs to be a 40 digit hex number')
+                return sys.exit(error_codes.numeric_status['AVOCADO_CRASH'])
 
         job_instance = job.Job(args)
         rc = job_instance.run()
@@ -185,30 +130,3 @@ class TestRunner(plugin.Plugin):
             self.parser.print_help()
 
         return rc
-
-
-class SystemInformation(plugin.Plugin):
-
-    """
-    Collect system information
-    """
-
-    name = 'sysinfo'
-    enabled = True
-
-    def configure(self, parser):
-        """
-        Add the subparser for the run action.
-
-        :param parser: Main test runner parser.
-        """
-        self.parser = parser.subcommands.add_parser(
-            'sysinfo',
-            help='Collect system information')
-        self.parser.add_argument('sysinfodir', type=str,
-                                 help='Dir where to dump sysinfo',
-                                 nargs='?', default='')
-        super(SystemInformation, self).configure(self.parser)
-
-    def run(self, args):
-        sysinfo.collect_sysinfo(args)
